+++ src/components/navigation-accordion/__tests__/navigation-accordion-test-cases.js
@@ -97,6 +97,54 @@ testCases.withThirdLevelItems = {
   }
 };
 
+testCases.withThirdLevelItemsWithIcons = {
+  description: 'Shows third level items with icons',
+  component: NavigationAccordion,
+  props: {
+    currentPath: 'page-one',
+    contents: {
+      firstLevelItems: [
+        {
+          title: 'Title one',
+          path: 'page-one'
+        },
+        {
+          title: 'Title two',
+          path: 'page-two'
+        }
+      ],
+      secondLevelItems: [
+        {
+          title: 'Heading one',
+          path: 'heading-one',
+          thirdLevelItems: [
+            {
+              title: 'Subheading one',
+              path: '',
+              icon: 'paint'
+            },
+            {
+              title: 'Subheading two',
+              path: 'subheading-two',
+              icon: 'paint'
+            },
+            {
+              title: 'Subheading three is long so we should wrap it nicely',
+              path: 'subheading-three',
+              icon: 'star'
+            }
+          ]
+        },
+        {
+          title: 'Heading two',
+          path: 'heading-two'
+        }
+      ]
+    },
+    onDropdownChange: () => {}
+  }
+};
+
 testCases.withTags = {
   description: 'Titles with tags',
   component: NavigationAccordion,